Refactor balloon index action.
No functional changes intended.

This is in preparation of replacing balloons.in, see #425.
